fix(MidashigoMode): handle unfixed roman letter "n" with successive capital consonants
Fix a bug in MidashigoMode where when an unfixed roman letter "n" is present and capital consonants are input in succession, the input wasn't handled correctly. After this fix, when such an input is received, "ん" is added to the word stem and MidashigoMode continues with the consonant as unfixed okuri-gana.